Create Jumble word
Prashant started to attend programming lessons. On the first lesson his task was to write a simple program. The program was supposed to do the following: in the given string, consisting if uppercase and lowercase Latin letters, it:

deletes all the vowels,
inserts a character "." before each consonant,
replaces all uppercase consonants with corresponding lowercase ones.
Vowels are letters "A", "O", "Y", "E", "U", "I", and the rest are consonants. The program's input is exactly one string, it should return the output as a single string, resulting after the program's processing the initial string.

Help Prashant cope with this easy task.

Input First line of input contains an integer T number of test cases. The next T line represents input string of Prashant's program. This string only consists of uppercase and lowercase Latin letters and its length is from 1 to 100, inclusive.

Output Print the resulting string. It is guaranteed that this string is not empty.