Create cyclic shift hackerearth
You are given a number  represented as a binary representation of  bits. You are also given a number  and a character .

Determine a number  that is generated after cyclically shifting the binary representation of  by  positions either left if  or right is .

Input format 

The first line contains an integer  representing the number of queries.
The next  lines contain  as mentioned in the problem statement.
Output format

Print  integers in a separate line representing the answer to each query.