Create play next match
Manish is captain of a cricket team. Before every match, he checks whether the players are mentally prepared or not. this time he opts for a numerical problem for checking the mental ability of his players. He gives a set of numbers, as a player you have to find the sum of 2(position of the first SET bit in number). solve this question if you want to play in next match.

Input:

The first line of input contains an integer T, where T is the number of test cases.
The first line of each test case contains an integer N.
The second line of each test case contains N, space separated number.

Output:

single line for each test case